expand-sequence: Update tests for recent changes
This should have been part of a previous commit.  So few parts of the system
utilize tests that I didn't notice that this was failing until after the
commit, since I'm not used to running them...even though I wrote them.

This system is being replaced with TAMER, so these changes largely just get
the tests working properly again.

DEV-15095