depgen: Remove lax symbol concept
I had to dig through the old repository (prior to extracting into this one)
to see why this was introduced.  It seems that it was for the linker, and
TAMER's linker has no concept of lax symbols, so this is not used.

To make matters worse, the code I modified here could not have worked (in
depgen) because `$syms/@name` _will always have a value_.

Anyway, removing this has no effect on the compiled packages.

DEV-15114