depgen (preproc:symtable): Simplify symbol iteration
This just continues to refactor to try to make sense of this code, which has
evolved into quite the mess over the years.  The two primary goals are
to (a) find possible optimizations and (b) make sure how this functions is
clear for when it's reimplemented in TAMER.

I'm doing this in small commits so that the steps are more obvious.

The specific list of attributes is what I found to be output in practice in
the `xmlo` files.

DEV-15114