depgen (preproc:symtable): Use for-each-group for deduplication
And here's the optimization that I had wanted to perform all along, but it
took some time to confidently get to this point.

`preceding-sibling` was used since the XSLT1 days for deduplication, before
`for-each-group` existed.  It works just fine for small inputs, but the
problem is we're doing this many thousands of times in larger packages, and
that really adds up.  (In this case, shaving ~8s off of one of our large
packages with ~20k symbols in play.)

DEV-15114