summary: Warn on missing packages during typesetting
This restores the previous behavior (from before a couple commits ago), but
adds a warning.

We have a GitLab pipeline internally where we break apart certain
directories into separate jobs to parallelize them.  There is a bug relating
to build and retaining artifacts for rate tables that only manifests itself
for Summary Pages, and that needs to be resolved before this is removed.

DEV-15153