tamer: diagnose::diagnostic_error_sum: New macro
This removes error sum type boilerplate, something that I've long wanted to
do.  It took me until just before pausing TAMER work to have confidence that
I've settled on a pattern for errors with the diagnostic system.

This commit was originally demonstrated as part of a WIP system; I
cherry-picked it and applied it instead to tamec for illustration.  More
will follow.