progtest: runner: Fail on missing input files
The runner works by concatenating all provided arguments into a single file,
which it then provides to the runner.  This is done via process substitution
rather than creating a temporary file.

This works fine, but process substitution (using `<()`) does not propagate
exit status, which means that `cat` failures did not cause the script to
fail.

Since progtest is invoked by an automated process based on the contents of
the filesystem, this wasn't seen until someone committed a test with spaces
in the filename; Make does not handle spaces in filenames, and so this
resulted in the constituent word-splitted components being passed to this
script, which of course did not exist.

DEV-16407