build-aux/Makefile.am (suppliers.mk): Append to package-dfns.dep rather than replace
(This is a change that belongs in liza-proguic, not TAME, but it has yet
to be moved.  I'd rather not hold up this change for that.)

Commit 6f67a4d introduced support for step-level packages in a rather
klugy way.  I do not thing it was my intent to fully overwrite this
file, as that makes no sense.

This will keep the dependencies discovered for package-dfns.xml and
augment it with the result of the filesystem scan.

Previously, replacing the dependencies meant that whether a UI package
built depended on whether some other package imported it.  This was
especially inconsistent with parallel builds, which casued races.
Furthermore, it means that some packages had stale state on rebuild,
which surely has caused some confusion and headaches over the years.