tamer: asg::graph: object_rel! and AsgRelMut adjustment for modern min_specialization
The `min_specialization` feature has evolved since this was originally
written: at some point in the past couple of years, the feature now
rejects this original implementation with (for example) this error:

```
error: specialization impl does not specialize any associated items
   --> src/asg/graph/object/rel.rs:182:13
    |
182 | /             impl AsgRelMut<$kind> for $from {
183 | |                 $( $($impl)* )?
184 | |             }
    | |_____________^
    |
   ::: src/asg/graph/object/tpl.rs:239:1
```

This seems to be because the `impl` is empty in the default case, with
the expectation of falling back to the default implementation.

My first attempt was to entirely omit an `impl` if there is no override
(in `object_rel`'s new `@trait_specialization`), however this gave a
different type of error elsewhere in the system, as if Rust only
recognized `AsgRelMut` for `Doc` (the only one that has its own
`pre_add_edge` override).  I have not yet investigated that further.

However, I don't object to this implementation, and I'm pretty sure I
had attempted it previously: the default implementation can panic, and
permitted object relationships can default to a simple commit.

The benefit of this approach is that, if there's a bug in the system
that permits an edge to be added when it shouldn't be (if the types fail
to enforce the relation statically), then the system will panic rather
than adding invalid data to the graph.  Given that feature, this may be
a desirable end state.