tamer: nightly: Upgrade 2023-04-15 => 2024-06-24
Note: _This commit will cause many warnings that will be fixed in the
commits that follow._  This commit already contains too many unrelated
changes.

This has been eating at me for some time, wondering how much work it
would take to upgrade this.  I'm surprised that this has been reduced to
such a small changeset.

This commit represents changes that are _not_ backwards-compatible, and
so have to be made in conjunction with a compiler version update; the
reason for that is that TAMER uses a number of unstable features that
are still under development and whose implementations change over time.
This is a calculated risk, for the benefits that they provide.

(A number of features _have_ been stabalized; see upcoming commit, or
warnings in this one.)

This commit, in particular, addresses:

  - `adt_const_params` now has a separate feature flag
    `unsized_const_params` which are needed for `&'static str` in const
    contexts.  This one had me worried at first, since it would have
    been very disappointing to have to roll that use back.
  - As noted in the (deleted) `rust-toolchain.toml` comment,
    `adt_const_prams` now requires `ConstParamTy` for types that are
    used in const contexts, so that was implemented for a couple of
    them.
    - But it could not be implemented for `NonZero*` types, since those
      types themselves do not yet implement that type.  Consequently,
      this creates a really awkward asymmetry between static strings and
      (the supposedly equivalent) `SymbolId` that I hope will eventually
      be resolved; in the meantime, `nir::literal` was converted into a
      partially applied function.  I haven't yet checked to see if Rust
      optimizes it all the same anyway, given that its value should be
      statically known.
  - `ParseState::Super` no longer uses `ClosedParseState`, beacuse the
    `Self` default was no longer accepted.  This is due to further
    development of the unstable `associated_type_defaults` feature.
    That trait bound was instead moved to `ParseState::delegate_child`.
    While unfortunate (it won't catch problems as early and it may be
    more confusing when it does catch them), it does save us from a ton
    of boilerplate (having to specify `Super` for every `ParseState`
    `impl` even though most do not have to care about such a feature),
    so I'm glad that worked out.  It might have resulted in a macro if I
    had to do that, to keep the concern encapsulated.
  - `Span` had some surprising unit test failures for the packed
    little-endian representation.  Fortunately, changing `repr(packed)`
    to `repr(C, packed)` fixed that issue.

There are some other miscellaneous changes documented throughout, but
the above covers the majority of them.