tamer: Fix miscellaneous warnings from nightly upgrade
Some clippy, some rustc.

These are warnings, not errors.  I'm particularly fond of the warning for the
static check for parsing the invalid UTF-8 byte sequence (in the test case):

```
warning: calls to `std::str::from_utf8` with an invalid literal always return an error
   --> src/sym/interner.rs:628:24
    |
624 |         let bytes = &[0b11000000u8];
    |                      -------------- the literal was valid UTF-8 up to the 0 bytes
...
628 |         match (result, from_utf8(bytes)) {
    |                        ^^^^^^^^^^^^^^^^
    |
    = note: `#[warn(invalid_from_utf8)]` on by default
```

There are more, but I wanted to separate these.