tamer: asg::graph::object::rel::ObjectRel::{narrows_into,widen}: Allow dead_code
These are currently used only by tests.  I want to keep them as part of
the API, because as others get into the system, I want them to clearly
see (using docs and their LSP) that this is a method available to them.

We need a careful balance between "this may occasionally become unused,
but it's important as an API", in a library sense, and "we want to
delete code that's not used".

Maybe this ought to just become a library crate.

Anyway, we'll continue to reason about that.