tamer: Canonical manual Clone impls for Copy types
This is a new lint (as of the nightly upgrade):

```
warning: non-canonical implementation of `clone` on a `Copy` type
    --> src/asg/graph/object.rs:1084:29
     |
1084 |       fn clone(&self) -> Self {
     |  _____________________________^
1085 | |         Self(self.0)
1086 | |     }
     | |_____^ help: change this to: `{ *self }`
     |
     = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#non_canonical_clone_impl
```

Ideally, we're not manually implementing `Clone` or `Copy` at all; the
issue at play here is that we have phantom types (which are used to
propagte types but do not have any associated data).  Since they have no
data, they have nothing to copy/clone.  However, `derive` adds the
`Copy` and `Clone` trait bounds to all other bounds, which would require
that the bounds utilized only by phantom types also implement
`Copy`/`Clone`, which is unnecessarily restrictive at this point.