tamer: xir::parse::ele::test: Indent tokens
TAMER's modern testing standard is to intent tokens in a way that aids
in understanding the structure of the test.

`ele_parse` currently has no documentation, apparently, and so that
makes clear test cases even more important.

If we're comparing with AIR, for example (e.g. `asg::air::expr::test`),
the next step ought to be rolling failurse up into a declarative
assertion as well, to further simplify the tests.